# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/shard-analyzer": {
            "post": {
                "description": "Endpoint to take cat/shards output and examine they are rightly sharded or not and recommend the right ones if necessary.",
                "consumes": [
                    "application/text"
                ],
                "produces": [
                    "application/json",
                    " application/pdf"
                ],
                "summary": "Recommend shard strategies",
                "operationId": "shardAnalyzerPost",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Cluster-name",
                        "description": "Cluster Name",
                        "name": "clusterName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "AWS Customer",
                        "description": "Customer Name",
                        "name": "customerName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 30,
                        "description": "Target Shard Size in GB",
                        "name": "targetShardSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 3,
                        "description": "Number of Azs for the cluster",
                        "name": "azs",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "default": false,
                        "description": "If log analytics, 1 replica is recommended. If not the replica count will be retained",
                        "name": "isSearchWorkload",
                        "in": "query"
                    },
                    {
                        "description": "Output of cat/shards.",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}